requirements.txt
# Django
django==4.2.17
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
django-import-export==4.3.3

#photo and images
pillow==11.1.0
requests==2.32.3
